Refactor: enforce coding conventions

+++ src/modules/schedule/services/schedule-date-override.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import {
   EntityManager,
-  NotFoundError,
   ForeignKeyConstraintViolationException,
 } from '@mikro-orm/postgresql';
 
@@ -10,8 +9,8 @@ import { isOverlapping } from '@/utils/helpers/time';
 import { Schedule } from '../entities/schedule.entity';
 import { CreateDateOverrideDto } from '../dto/create-date-override.dto';
 import { UpdateDateOverrideDto } from '../dto/update-date-override.dto';
+import { ScheduleDateOverrideResDto } from '../dto/date-override-res.dto';
 import { ScheduleDateOverride } from '../entities/schedule-date-override.entity';
-import { ScheduleDateOverrideResponseDto } from '../dto/date-override-response.dto';
 import { ScheduleNotFoundException } from '../exceptions/schedule-not-found.exception';
 import { OverlappingIntervalsException } from '../exceptions/overlapping-intervals.exception';
 import { DateOverrideNotFoundException } from '../exceptions/date-override-not-found.exception';
@@ -19,27 +18,30 @@ import { DateOverrideNotFoundException } from '../exceptions/date-override-not-f
 @Injectable()
 export class ScheduleDateOverrideService {
   constructor(private em: EntityManager) {}
-  async create(dto: {
+  async create({
+    scheduleId,
+    dateOverrideData,
+  }: {
     scheduleId: number;
     dateOverrideData: CreateDateOverrideDto;
   }) {
-    const { intervals } = dto.dateOverrideData;
+    const { intervals } = dateOverrideData;
 
-    if (intervals.length > 1) {
-      intervals.forEach((interval, index) => {
-        const nextIntervals = intervals.slice(index + 1);
-        nextIntervals.forEach((nextInterval) => {
-          if (isOverlapping(interval, nextInterval)) {
-            throw new OverlappingIntervalsException(interval, nextInterval);
-          }
-        });
-      });
+    for (const [idx, currentInterval] of intervals.entries()) {
+      for (const comparingInterval of intervals.slice(idx + 1)) {
+        if (isOverlapping(currentInterval, comparingInterval)) {
+          throw new OverlappingIntervalsException(
+            currentInterval,
+            comparingInterval,
+          );
+        }
+      }
     }
 
     try {
       const existingOverrides = await this.em.find(ScheduleDateOverride, {
-        schedule: { id: dto.scheduleId },
-        date: dto.dateOverrideData.date,
+        schedule: { id: scheduleId },
+        date: dateOverrideData.date,
       });
 
       if (existingOverrides.length > 0) {
@@ -48,63 +50,71 @@ export class ScheduleDateOverrideService {
 
       const newOverrides = intervals.map((interval) =>
         this.em.create(ScheduleDateOverride, {
-          date: new Date(dto.dateOverrideData.date),
-          startTime: interval.startTime,
           endTime: interval.endTime,
-          schedule: this.em.getReference(Schedule, dto.scheduleId),
+          startTime: interval.startTime,
+          date: new Date(dateOverrideData.date),
+          schedule: this.em.getReference(Schedule, scheduleId),
         }),
       );
 
       await this.em.flush();
 
       return newOverrides.map((override) =>
-        ScheduleDateOverrideResponseDto.fromEntity(override),
+        ScheduleDateOverrideResDto.fromEntity(override),
       );
     } catch (error) {
       if (error instanceof ForeignKeyConstraintViolationException) {
-        throw new ScheduleNotFoundException(dto.scheduleId);
+        throw new ScheduleNotFoundException(scheduleId);
       }
+
+      throw error;
     }
   }
 
-  async update(dto: {
+  async update({
+    scheduleId,
+    dateOverrideId,
+    dateOverrideData,
+  }: {
     scheduleId: number;
     dateOverrideId: number;
     dateOverrideData: UpdateDateOverrideDto;
-  }) {
-    try {
-      const dateOverride = await this.em.findOneOrFail(ScheduleDateOverride, {
-        id: dto.dateOverrideId,
-        schedule: { id: dto.scheduleId },
-      });
+  }): Promise<ScheduleDateOverrideResDto> {
+    const dateOverride = await this.em.findOne(ScheduleDateOverride, {
+      id: dateOverrideId,
+      schedule: { id: scheduleId },
+    });
 
-      this.em.assign(dateOverride, {
-        startTime: dto.dateOverrideData.startTime,
-        endTime: dto.dateOverrideData.endTime,
-      });
-
-      await this.em.flush();
-
-      return ScheduleDateOverrideResponseDto.fromEntity(dateOverride);
-    } catch (error) {
-      if (error instanceof NotFoundError) {
-        throw new DateOverrideNotFoundException(dto.dateOverrideId);
-      }
+    if (!dateOverride) {
+      throw new DateOverrideNotFoundException(dateOverrideId);
     }
+
+    this.em.assign(dateOverride, {
+      startTime: dateOverrideData.startTime,
+      endTime: dateOverrideData.endTime,
+    });
+
+    await this.em.flush();
+
+    return ScheduleDateOverrideResDto.fromEntity(dateOverride);
   }
 
-  async delete(dto: { scheduleId: number; dateOverrideId: number }) {
-    try {
-      const dateOverride = await this.em.findOneOrFail(ScheduleDateOverride, {
-        id: dto.dateOverrideId,
-        schedule: { id: dto.scheduleId },
-      });
+  async delete({
+    scheduleId,
+    dateOverrideId,
+  }: {
+    scheduleId: number;
+    dateOverrideId: number;
+  }): Promise<void> {
+    const dateOverride = await this.em.findOne(ScheduleDateOverride, {
+      id: dateOverrideId,
+      schedule: { id: scheduleId },
+    });
 
-      await this.em.removeAndFlush(dateOverride);
-    } catch (error) {
-      if (error instanceof NotFoundError) {
-        throw new DateOverrideNotFoundException(dto.dateOverrideId);
-      }
+    if (!dateOverride) {
+      throw new DateOverrideNotFoundException(dateOverrideId);
     }
+
+    await this.em.removeAndFlush(dateOverride);
   }
 }
